Organize Python package
