+++ python/setup.cfg
@@ -1,0 +1,29 @@
+[metadata]
+
+name = SMIXS
+version = 0.1.0
+author = Peter Mlakar
+author_email = ADD_EMAIL_HERE
+description = SMIXS smoothing spline clustering algorithm for longitudinal data analysis
+long_description = SMIXS smoothing spline clustering algorithm for longitudinal data analysis
+long_description_content_type = text/markdown
+classifiers =
+    Development Status :: 2 - Pre-Alpha
+    Intended Audience :: Developers
+    Topic :: Software Development :: Localization
+    License :: OSI Approved :: MIT License
+    Programming Language :: Python :: 3
+
+project_urls =
+    source code = https://github.com/MysteriousMan231/SMIXS
+    bug tracker = https://github.com/MysteriousMan231/SMIXS/issues
+
+[options]
+package_dir =
+    = .
+packages = find:
+install_requires =
+    numpy
+    scikit-learn
+
+python_requires = >=3.8
